requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
black==20.8b1
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cloudevents==1.2.0
coverage==5.4
coverage-badge==1.0.1
cryptography==3.4.6
deprecation==2.1.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
functions-framework==2.1.2
google-api-core==1.26.1
google-auth==1.24.0
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-cloud-secret-manager==2.2.0
google-cloud-tasks==2.2.0
google-crc32c==1.1.2
google-oauth==1.0.1
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
gunicorn==20.0.4
identify==2.1.1
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
parameterized==0.8.1
pathspec==0.8.1
pre-commit==2.10.1
proto-plus==1.14.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.23.0
requests-toolbelt==0.9.1
rollbar==0.14.7
rsa==4.7.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
virtualenv==20.4.2
watchdog==1.0.2
Werkzeug==1.0.1
